+++ Produit.h
@@ -15,19 +15,19 @@ private:
 
 public:
 
-	Produit(std::string _produit, double _id, std::string _description, double _qtite, float _prix);
+	Produit(std::string _produit, int _id, std::string _description, int _qtite, float _prix);
 
 
 		void setProduit(std::string _produit);
 		void setDescription(std::string _description);
-		void setId(double _id);
-		void setQtite(double _qtite);
+		void setId(int _id);
+		void setQtite(int _qtite);
 		void setPrix(float _prix);
 
 		std::string getNameProduit() const;
 		std::string getDescription() const;
-		double getId() const;
-		double getQtite() const;
+		int getId() const;
+		int getQtite() const;
 		float getPrix() const;
